% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{default_style}
\alias{default_style}
\title{Default \code{prettycode} syntax highlighting style}
\usage{
default_style()
}
\description{
A style function must return a named list of functions. Possible
entries:
\itemize{
\item \code{reserved}: reserved words
\item \code{number}: numeric literals
\item \code{null}: the \code{NULL} constant
\item \code{operator}: operators, including assignment
\item \code{call}: function calls
\item \code{string}: character literals
\item \code{comment}: comments
}
}
\details{
Each entry in a list must be a function that takes a character
scalar, and returns a character scalar. The default style adds
ANSI formatting to the code.

Note that you can also change the code if you like, e.g. to include
a unicode arrow character instead of the two-character assignment
operator.
}
\examples{
highlight(deparse(get), style = default_style())
}

