% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Syntax highlight R code}
\usage{
highlight(code, style = default_style())
}
\arguments{
\item{code}{Character vector, each element is one line of code.}

\item{style}{Style functions, see \code{\link[=default_style]{default_style()}}.}
}
\value{
Character vector, the highlighted code.
}
\description{
Syntax highlight R code
}
\examples{
highlight(deparse(ls))
cat(highlight(deparse(ls)), sep = "\\n")
}
