% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettycode-package.R
\docType{package}
\name{prettycode-package}
\alias{prettycode-package}
\title{prettycode: Pretty Print R Code in the Terminal}
\description{
Replace the standard print method for functions with one that performs syntax highlighting, using ANSI colors, if the terminal supports them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/prettycode#readme}
  \item \url{https://r-lib.github.io/prettycode/}
  \item Report bugs at \url{https://github.com/r-lib/prettycode/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
