% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{prettycode}
\alias{prettycode}
\title{Pretty Print R Code in the Terminal}
\usage{
prettycode()
}
\description{
Replace the standard print method for functions with one that performs
syntax highlighting, using ANSI colors, if the terminal supports them.
}
