% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.function}
\alias{print.function}
\title{Print a function with syntax highlighting}
\usage{
\method{print}{`function`}(x, useSource = TRUE, style = default_style(), ...)
}
\arguments{
\item{x}{Function to print.}

\item{useSource}{Whether to use the stored source code, if available.}

\item{style}{The highlight style to use, see \code{\link[=default_style]{default_style()}}.}

\item{...}{Not used currently, for compatibility with the \code{print}
generic.}
}
\value{
The function, invisibly.
}
\description{
To turn on pretty printing of functions, you need to call
\code{prettycode::prettycode()}. It might be a good idea to call it
from your \code{.Rprofile}.
}
