% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prettycode}
\alias{prettycode}
\alias{prettycode-package}
\title{Pretty Print R Code in the Terminal}
\description{
Replace the standard print method for functions with one that performs
syntax highlighting, using ANSI colors, if the terminal supports them.
}

